Validate complete registered profiles and organize protocol tests
